Added the functions to find the best-fit operators
It contains a SVD and a Leastsqr function to replace the numpy ones.
There are the DMD, DMD with control with B known, B unknown, and Operator inference code. It only gets the operators of the system.
The "matlab" codes give the modes, the eigenvalues, etc ... for the Dynamic Mode Decomposition.
The rest of the script needed to "print" the results can be taken in the tutorial_opinf python file.